fix 32-bit overflow in `mi_page_start` for pages ≥ 4 GiB after their page meta info (issue #1309)

+++ include/mimalloc/internal.h
@@ -683,7 +683,8 @@ static inline size_t mi_page_block_size(const mi_page_t* page) {
 
 // Page start
 static inline uint8_t* mi_page_start(const mi_page_t* page) {
-  return (uint8_t*)page + (page->page_woffset * MI_SIZE_SIZE);
+  // multiplication must be done in `size_t`; in a 32-bit multiplication the offset wraps for pages whose blocks start 4 GiB or more after the page meta info
+  return (uint8_t*)page + ((size_t)page->page_woffset * MI_SIZE_SIZE);
 }
 
 static inline size_t mi_page_size(const mi_page_t* page) {
